middleware/requirements.txt
fastapi
uvicorn
requests
python-socketio[client]
python-multipart
aiohttp
